fixed wav length

+++ MAX/Game/Vendors/cocos2d-2.0-x-2.0.4/CocosDenshion/ios/SimpleAudioEngine.mm
@@ -242,7 +242,7 @@ unsigned int SimpleAudioEngine::playEffect(const char* pszFilePath, bool bLoop)
             return static_playEffect(pszFilePath, bLoop);
 }
     
-unsigned int SimpleAudioEngine::lengthOfEffect(const char* pszFilePath)
+float SimpleAudioEngine::lengthOfEffect(const char* pszFilePath)
 {
     return static_lengthOfEffect(pszFilePath);
 }
